+++ src/client/game/Game.ts
@@ -1,16 +1,22 @@
-// Code from https://www.babylonjs-playground.com/#3EDS3A#22
 
 import * as Babylon from "babylonjs";
 import { Vector3 } from "babylonjs";
 
 import { GUIManager, IGUIManager } from "@client/gui";
-import { ClusterClient, IClusterClient } from "@client/cluster";
 import { IView } from "@client/view";
 
 import { AssetManager, IAssetManager } from "@client/assets";
-import { PlayerMotor, IPlayerMotor } from "@client/movement";
+import { PlayerMotor } from "@client/movement";
 
-import { ClusterData, IClusterData } from "@share/data/cluster-data";
+import { IClusterData } from "@share/data/cluster-data";
+import { ClientIncoming, ClientOutgoing } from "@client/socket";
+import { Socket } from "socket.io-client";
+import { Action } from "@share/action/Action";
+import { ClientActionProcessor } from "@client/action/ClientActionProcessor";
+import { LoadClusterAction } from "@share/action/cluster/LoadClusterAction";
+import { UnloadClusterAction } from "@share/action/cluster/UnloadClusterAction";
+import { IClusterManager } from "@client/cluster/IClusterManager";
+import { ClusterManager } from "@client/cluster";
 
 /**
  * The runner class for all game logic.
@@ -21,16 +27,20 @@ export class Game {
   private _engine: Babylon.Engine;
   private _scene: Babylon.Scene;
 
+  // Networking
+  private _outgoing: ClientOutgoing;
+  private _actionProcessor: ClientActionProcessor;
+
   // Lighting
   private _sun: Babylon.DirectionalLight;
   private _hemisphericLight: Babylon.HemisphericLight;
   private _shadowGenerator: Babylon.ShadowGenerator | null = null;
 
   // Game elements
-  private _cluster: IClusterClient;
-  private _motor: IPlayerMotor;
+  private _clusterManager: IClusterManager | null = null;
+  private _motor: PlayerMotor | null = null;
   private _gui: IGUIManager;
-  private _assetManager: IAssetManager | null = null;
+  private _assetManager: IAssetManager;
 
   /**
    * Create a new Game.
@@ -39,7 +49,7 @@ export class Game {
    */
   constructor(
     private _view: IView,
-    clusterString: string,
+    socket: Socket,
     debugMode = false
   )
   {
@@ -64,15 +74,10 @@ export class Game {
     this._shadowGenerator = new Babylon.ShadowGenerator(1024, this._sun);
     this._shadowGenerator.usePoissonSampling = true;
 
-    // Get world data from the server
-    const clusterData = ClusterData.fromStringRep(clusterString);
-
-    this._cluster = new ClusterClient(clusterData, this._shadowGenerator, this._assetManager);
-    this._cluster.remesh();
-
-    // Create local player motor
-    this._motor = new PlayerMotor(
-      _view.getCanvas(), this._engine, this._scene, this._cluster, new Vector3(50, 70, -50), true);
+    // Init communications
+    this._outgoing = new ClientOutgoing(socket);
+    this._actionProcessor = new ClientActionProcessor(this, this._outgoing);
+    new ClientIncoming(socket, this._actionProcessor);
 
     this._gui = new GUIManager();
     this._gui.mainMenuGui();
@@ -83,15 +88,65 @@ export class Game {
     const scene = this._scene;
     const engine = this._engine;
     this._engine.runRenderLoop(() => {
-      scene.render();
+      if (this._motor !== null)
+        scene.render();
       fpsElement.innerHTML = engine.getFps().toFixed();
     });
   }
 
   /**
+   * Process an Action to update the game state.
+   */
+  public processAction(action: Action) {
+    
+    // TODO: Move cluster loading/unloading inside ClusterManager
+    if (action instanceof LoadClusterAction) {
+      this.loadCluster(action.cluster);
+    }
+    else if (action instanceof UnloadClusterAction) {
+      this.unloadCluster();
+    }
+
+    this._clusterManager?.processAction(action);
+  }
+
+  /**
+   * Load a cluster from a ClusterData.
+   */
+  public loadCluster(cluster: IClusterData) {
+    this.unloadCluster();
+
+    // Initialize cluster client
+    this._clusterManager = new ClusterManager(cluster, this._shadowGenerator, this._assetManager);
+    this._clusterManager.remesh();
+
+    // Create local player motor
+    this._motor = new PlayerMotor(
+      this._view.getCanvas(), this._engine, this._scene, this._actionProcessor, new Vector3(50, 70, -50), true);
+  }
+
+  /**
+   * Remove the current cluster from the world.
+   */
+  public unloadCluster() {
+    if (this._clusterManager !== null) {
+      this._clusterManager.dispose();
+      this._clusterManager = null;
+    }
+  }
+
+  /**
+   * Get the IClusterManager.
+   * @returns This Game's IClusterManager instance.
+   */
+  public getCluster(): IClusterManager | null {
+    return this._clusterManager;
+  }
+
+  /**
    * Initialize the Babylon scene before adding objects.
    */
-  _initScene(debugMode: boolean = false): Babylon.Scene {
+  private _initScene(debugMode: boolean = false): Babylon.Scene {
 
     // Set up scene
     const scene = new Babylon.Scene(this._engine);
@@ -107,7 +162,7 @@ export class Game {
   /**
    * Add event listeners.
    */ 
-  _addEventListeners() {
+  private _addEventListeners() {
     const engine = this._engine;
 
     // Window resize listener