+++ src/client/movement/IPlayerMotor.ts
@@ -1,6 +1,0 @@
-/**
- * Control the local player movement.
- */
-export interface IPlayerMotor {
-  
-}