+++ src/client/movement/index.ts
@@ -1,2 +1,1 @@
 export { PlayerMotor } from "./PlayerMotor";
-export { IPlayerMotor } from "./IPlayerMotor";