+++ src/client/socket/index.ts
@@ -1,0 +1,2 @@
+export { ClientIncoming } from "./ClientIncoming";
+export { ClientOutgoing } from "./ClientOutgoing";