+++ src/server/game-server/GameServer.ts
@@ -1,33 +1,35 @@
 import { ClusterData, IClusterData } from "@share/data/cluster-data";
 import { ContinuousClusterGenerator } from "@server/cluster-gen";
-import { Outgoing } from "@server/socket";
 import { Socket } from "socket.io";
-import { Grid } from "@share/data/grid";
-import { Vector3 } from "babylonjs";
-import { SerializableGrid } from "@share/data/grid/SerializableGrid";
-import { ChunkCoordinate } from "@share/data/coordinate";
-import { join } from "path";
-import { writeFileSync } from "fs";
+import { Server as IOServer } from 'socket.io';
+import { Action } from "@share/action";
+import { ServerIncoming, ServerOutgoing } from "@server/socket";
+import { LoadClusterAction } from "@share/action/cluster/LoadClusterAction";
+import { AddBlockAction } from "@share/action/block/AddBlockAction";
+import { RemoveBlockAction } from "@share/action/block/RemoveBlockAction";
+import { ActionData } from "@server/action/ActionData";
 
 /**
  * Handler for server-side game logic.
  */
 export class GameServer {
   private cluster: IClusterData;
+  private outgoing: ServerOutgoing;
 
   /**
    * Initialize the game server.
    */
-  constructor(private readonly outgoing: Outgoing) {
+  constructor(
+    private readonly io: IOServer
+  ) {
+    // Set up server-client comms
+    this.outgoing = new ServerOutgoing(io);
+    new ServerIncoming(this, io);
 
     // Generate a cluster
     console.log("Generating world");
     this.cluster = ContinuousClusterGenerator.createWorldCluster();
     console.log("Done generating world");
-
-    const rep = this.cluster.toStringRep();
-    //writeFileSync(join(__dirname, "output.txt"), rep, {flag: 'w'});
-    const cluster2 = ClusterData.fromStringRep(rep);
   }
 
   /**
@@ -43,6 +45,22 @@ export class GameServer {
    * @param socket The socket of the user that connected.
    */
   onConnection(socket: Socket) {
-    this.outgoing.sendWorld(socket, this.cluster);
+    const action = new LoadClusterAction(this.cluster);
+    this.outgoing.sendAction(action);
+  }
+
+  /**
+   * Handle an action on the server side.
+   */
+  processAction(actionData: ActionData) {
+
+    const action = actionData.action;
+    this.cluster.processAction(action);
+
+    if (action instanceof AddBlockAction || action instanceof RemoveBlockAction) {
+
+      // Update all clients except for the one that sent the action
+      this.outgoing.sendActionFrom(action, actionData.actor);
+    }
   }
 }