+++ src/share/data/cluster-data/IClusterData.ts
@@ -3,6 +3,7 @@ import { IChunkData } from "./chunk-data/IChunkData";
 import { Vector3 } from "babylonjs";
 import { IChunkCoordinate } from "@share/data/coordinate/chunk-coordinate";
 import { IAbsoluteCoordinate } from "@share/data/coordinate/absolute-coordinate";
+import { Action } from "@share/action";
 
 /**
  * Manage multiple chunks, generating their meshes and loading them
@@ -43,6 +44,12 @@ export interface IClusterData extends Iterable<[IChunkCoordinate, IChunkData]> {
    *   in this cluster.
    */
   toStringRep(): string;
+
+  /**
+   * Apply an Action to this cluster. If the action is not
+   * applicable, do nothing.
+   */
+  processAction(action: Action): void;
 
   /**
    * Get iterator for all chunks in the cluster and