+++ src/share/data/coordinate/chunk-coordinate/ChunkCoordinate.ts
@@ -8,7 +8,6 @@ import { Settings } from "@share/config/Settings";
  * Chunk coordinates. An increment of 1 shifts
  * to an adjacent chunk.
  * @throws {TypeError} If decimals are used as an input.
- * @deprecated Use Vector3 instead
  */
 export class ChunkCoordinate implements IChunkCoordinate {