+++ src/share/data/coordinate/chunk-coordinate/IChunkCoordinate.ts
@@ -6,7 +6,6 @@ import { ICoordinate } from "../ICoordinate";
  * Interface for chunk coordinates. An increment of 1 shifts
  * to an adjacent chunk.
  * @throws {TypeError} If decimals are used as an input.
- * @deprecated Use Vector3 instead
  */
 export interface IChunkCoordinate extends ICoordinate {
   get x(): number;