+++ src/share/data/coordinate/relative-coordinate/RelativeCoordinate.ts
@@ -1,21 +1,23 @@
 import { IRelativeCoordinate } from ".";
 import { AbsoluteCoordinate, IAbsoluteCoordinate } from "../absolute-coordinate";
-import { IChunkCoordinate } from "../chunk-coordinate";
+import { ChunkCoordinate, IChunkCoordinate } from "../chunk-coordinate";
 import { Vector3 } from "babylonjs";
 import { Settings } from "@share/config/Settings";
 
 /**
  * Interface for a relative coordinate, which stores the
  * position of a block relative to its chunk.
- * 
- * The x, y and z coordinates must be within 0 and the chunk size minus 1.
- * @throws {RangeError} If the x, y or z coordinate is not within 0 and the chunk size minus 1.
- * @throws {TypeError} If decimals are used as an input.
- * @deprecated Use Vector3 instead
  */
 
 export class RelativeCoordinate implements IRelativeCoordinate {
   
+  /**
+   * Create a new RelativeCoordinate.
+   * 
+   * The x, y and z coordinates must be within 0 and the chunk size minus 1.
+   * @throws {RangeError} If the x, y or z coordinate is not within 0 and the chunk size minus 1.
+   * @throws {TypeError} If decimals are used as an input.
+   */
   constructor(
     public readonly x: number,
     public readonly y: number,
@@ -36,6 +38,35 @@ export class RelativeCoordinate implements IRelativeCoordinate {
     if (x % 1 != 0 || y % 1 != 0 || z % 1 != 0) {
       throw new TypeError(`Cannot use decimals in RelativeCoordinate initialization - got ${x} ${y} ${z}`);
     }
+  }
+
+  /**
+   * Create a RelativeCoordinate from a string representation.
+   * @param rep The string representation of the coordinate.
+   * @returns A new RelativeCoordinate with values populated from rep.
+   */
+  static fromString(rep: string): RelativeCoordinate {
+
+    let inner = rep.substring(1, rep.length - 1);
+    let split = inner.split(",");
+
+    let x = parseInt(split[0]);
+    let y = parseInt(split[1]);
+    let z = parseInt(split[2]);
+    let cx = parseInt(split[3]);
+    let cy = parseInt(split[4]);
+    let cz = parseInt(split[5]);
+
+    return new RelativeCoordinate(x, y, z, new ChunkCoordinate(cx, cy, cz));
+  }
+
+  /**
+   * Get a string representation of this relative coordinate.
+   * @returns String representation of this relative coordinate.
+   */
+  toString(): string {
+    return "(" + this.x + "," + this.y + "," + this.z + "," +
+      this.chunkCoordinate.x + "," + this.chunkCoordinate.y + "," + this.chunkCoordinate.z + ")";
   }
 
   /**